% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC.R
\name{CC}
\alias{CC}
\title{Mortality curve calculation}
\usage{
CC(
  x,
  par,
  curve = c("gompertz", "makeham", "oppermann", "thiele", "wittsteinbumsted", "perks",
    "weibull", "vandermaen", "beard", "heligmanpollard", "rogersplanck", "siler",
    "martinelle", "thatcher")
)
}
\arguments{
\item{x}{vector of ages.}

\item{par}{vector of mortality curve parameter values.}

\item{curve}{name of mortality curve (including gompertz, makeham, oppermann, thiele, wittsteinbumsted, perks, weibull, vandermaen, beard, heligmanpollard, rogersplanck, siler, martinelle, thatcher).}
}
\value{
Calculated mortality rates based on selected mortality curve and parameter values.
}
\description{
Calculates mortality rates given mortality curve parameter values.
}
\details{
See \code{MC()} for more details of different mortality curves.
}
\examples{
CC(x=60:89,par=c(0.0000082,0.10771),curve="gompertz")

}
