% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curveS3.R
\name{curveS3}
\alias{curveS3}
\title{Common S3 methods for mortality curves}
\description{
Common S3 methods for objects returned by \code{MC()}:

coef(object) gives estimated parameter values.

fitted(object) gives fitted values.

predict(object,x) gives estimated mortality rate(s) for given age(s) x.

plot(object) plots data and fitted mortality curve.

deviance(object) gives weighted sum of squared residuals on \eqn{ln(m_x)}.

residuals(object) gives residuals on \eqn{ln(m_x)}.
}
