% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelS3.R
\name{modelS3}
\alias{modelS3}
\title{Common S3 methods for stochastic mortality models}
\description{
Common S3 methods for objects returned by \code{LCS()}, \code{RHS()}, \code{APCS()}, \code{CBDS()}, \code{CBDCS()}, \code{CBDQCS()}, and \code{STARS()}:

coef(object) gives estimated parameter values.

forecast::forecast(object) gives mortality forecasts smoothed by selected mortality curve.

plot(object) plots estimated parameter values, standardised residuals heatmap, data, and mortality forecasts smoothed by selected mortality curve.

residuals(object) gives standardised residuals on \eqn{ln(m_{x,t})}.
}
